% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSE_classifier.R
\name{TSE_classify}
\alias{TSE_classify}
\title{Classify Samples Based on TSE Categories}
\usage{
TSE_classify(x, minCor = 0.2)
}
\arguments{
\item{x}{A numeric matrix of normalized gene expression data (rows: genes, columns: samples).}

\item{minCor}{Minimum correlation required to assign a TSE category. Default is 0.2.}
}
\value{
A data frame with columns:
\describe{
\item{\code{TSE_category}}{Assigned TSE category for each sample.}
\item{\code{cor_pval}}{p-value of the correlation with the nearest centroid.}
\item{\code{separationLevel}}{A measure of how distinct the sample is from other categories.}
\item{\code{TSE_positive, TSE_neutral, TSE_negative}}{Correlation scores for each TSE category.}
\item{\code{sampleId}}{The ID of the sample.}
}
}
\description{
This function classifies samples into TSE categories ("TSE_positive", "TSE_neutral", "TSE_negative") based on correlation with predefined centroids.
}
\examples{
# Example usage with mock data
load(system.file("extdata", "centroids_TSE.RData", package = "TSECalculatoR"))
count_matrix <- matrix(rnorm(nrow(centroids_TSE * 10), mean = 1, sd = 1), nrow = nrow(centroids_TSE), ncol = 10)
rownames(count_matrix) <- rownames(centroids_TSE)
colnames(count_matrix) <- paste0("Sample", 1:10)
TSE_classify(count_matrix)
}
